srcs/utilsのリファクタリングとノーム対応
